Migration file modified

+++ database/migrations/2026_04_22_000001_rename_rent_due_date_to_rent_due_count_in_branches_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
         });
 
         Schema::table('branches', function (Blueprint $table) {
-            $table->unsignedTinyInteger('rent_due_count')->nullable()->change();
+            $table->unsignedBigInteger('rent_due_count')->nullable()->change();
         });
     }
